+++ src/cli/command/mod.rs
@@ -32,6 +32,8 @@ pub mod list;
 pub mod login;
 pub mod operator;
 pub mod operator_cloud_target_connect;
+#[cfg(not(feature = "sync_handler"))]
+pub mod operator_cloud_target_rekey;
 pub mod operator_init;
 #[cfg(not(feature = "sync_handler"))]
 pub mod operator_backup;